Rename RampRegistration Context and Provider to Ramp
