--- client/src/contexts/RampRegistration/index.ts
@@ -1,2 +1,0 @@
-export { default as RampRegistrationContext } from './RampRegistrationContext'
-export { default as RampRegistrationProvider } from './RampRegistrationProvider'
