+++ client/src/contexts/Registration/index.ts
@@ -1,0 +1,2 @@
+export { default as RegistrationContext } from './RegistrationContext'
+export { default as RegistrationProvider } from './RegistrationProvider'
