--- client/src/hooks/useRampRegistration.ts
@@ -1,9 +1,0 @@
-import { useContext } from 'react'
-
-import { RampRegistrationContext } from '../contexts/RampRegistration'
-
-const useRampRegistration = () => {
-  return { ...useContext(RampRegistrationContext) }
-}
-
-export default useRampRegistration
