+++ client/src/hooks/useRegistration.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react'
+
+import { RegistrationContext } from '../contexts/Registration'
+
+const useRampRegistration = () => {
+  return { ...useContext(RegistrationContext) }
+}
+
+export default useRampRegistration
